AG-17366-full-colmodel-rewrite
