+++ packages/ag-grid-community/src/main-internal.ts
@@ -337,7 +337,7 @@ export { BaseSelectionService } from './selection/baseSelectionService';
 export type { RowRangeSelectionContext } from './selection/rowRangeSelectionContext';
 export type { RowNodeSorter } from './sort/rowNodeSorter';
 export type { SortService } from './sort/sortService';
-export { _getSortModel } from './sort/sortUtils';
+export { _getSortModel } from './sort/sortService';
 export type { CellStyleService } from './styling/cellStyleService';
 export { coreDefaults as _coreThemeDefaults } from './theming/core/core-css';
 export {
